Added a new Native Stack navigator
The new Native navigator is now the top level navigator. When
you try to navigate, the event will bubble upwards until a
navigator can handle the event. In this case, if you navigate to
something which is not in the current navigator, it will bubble up.
Hence, you can now on demand either use the drawer or stack
navigator.