Wrok in progress. UI overhaul
Since i was implementing functionality to the old screens, i thought i might rewamp their design while im at it. Not done yet, but i thought a transition to a modal would be nice, since then the window in question does not need to have its own screen. Changes quite a lot but i believe this change is positive.