Fixes to model
Previously, localstorage was used as opposed to async storage. Local storage is instant and provided by the browsers and as such has near instant access time. Becauase of this, the module would constantly check the value to avoid uneccesary updates. However, async has significantly more overhead and this checking would quickly add up and slow down the device, This fixes that issue